+++ scripts/cordex_cv/__init__.py
@@ -1,0 +1,3 @@
+from .tools import create_cv
+
+__all__ = ["create_cv"]
