+++ scripts/cordex_cv/cli.py
@@ -1,0 +1,6 @@
+from .tools import create_cv as cv
+
+
+def create_cv():
+    print("creating CV")
+    cv()
